Added lexicon to notebook
The new features from the lexicon where added to the model selection
notebook. Changes were made to the attributes adder and lexicon manager
to be able to work in the notebook. Now the attributes adder class
returns the new calculated features instead of the whole received array
+ new features, and the lexicon manager returns 0.5 instead of None when
no token is present in the lexicon.